Add Kubernetes dashboard to Grafana
Like any other Grafana dashboard this dashboard is manually
created using the Grafana UI itself. It is then exported to
JSON using Grafana's "export" functionality. At last, the
format-dashboards.py script is used on the JSON file to
format all the dashboards in the same way, and increment the
version number each time a dashboard is updated.

Change-Id: Ib2abcb53b5b116633490730186d4a5157c91c8cd