Set correct host name in metrics
This commit fixes a bug where hostnames were not correct in
metrics collected by Snap and Hindsight. It relies on
Kubernetes' downward API and the spec.nodeName field. The
latter is only supported by Kubernetes 1.4 and higher, and
the deployment of stacklight-collector pods will fail if
Kubernetes 1.3 or lower is used.

Change-Id: I73cd35803a2201a09144bf925753156e47489cff
Depends-On: I293bb3aa113883c02f2e738f9d74291bf2f23d95
Partial-Bug: #1614484