Use standard way for building Heka Snap plugin
This commit changes the way we build the Heka publisher
plugin [1] in the Snap Docker image. The process of building
this plugin has changed in Snap, and is now much simpler.

This commit also fixes the build of the Snap image, which
has been failing in the 3rd-party CI periodic build job [2].

[1] <https://github.com/intelsdi-x/snap-plugin-publisher-heka>
[2] <https://jenkins-tp.ng.mirantis.net/view/build/job/mcp-build-images-build-number/24/console>

Change-Id: I1f63bfc52d52d5101ae4610c357b9e061935628c