Make the Heka build less fragile
Currently the Heka image build sometimes fails with the
following error:

	2016-09-07 18:08:09.219 2868 DEBUG fuel_ccp.build [-] heka: [91mCMake Error at /tmp/heka/cmake/message_proto.cmake:2 (message):
    Google protocol buffers 'protoc' must be installed, message.proto has been modified and needs to be regenerated.

This error is related to the "make" process trying to
re-generate "message.pb.go" using "protoc" which is not
present. And re-generating "message.pb.go" is not necessary
and shouldn't be done.

This commit attempts to fix the issue by touch'ing the
"message.pb.go" file to make sure the "make" process will
not attempt to re-generate it.

Change-Id: I91f12a99c813be99ba5a24ff65ca786eb97fea0c