Add Lua code for alarming
This commits adds Lua code for generating AFD (Anomaly and
Fault Detection) metrics based on the evaluation of alarms.
The Lua code was copied from the lma_collector Fuel plugin
[*], with changes to accomodate Hindsight and the versions
of lua_sandbox and lua_sandbox_extensions we rely on.

In the future we plan to move this Lua code in its own Git
repository. And the Hindsight Dockerfile will install the
Lua code in the image using Debian packages.

The cpu_alarms.lua and hindsight_cpu_alarms.cfg.j2 file will
be removed. Instead the operator will configure alarms
through a YAML file, and we will use a sidecar container for
generating Lua tables including alarm definitions and the
corresponding plugin configuration files.

[*] https://github.com/openstack/fuel-plugin-lma-collector/

Change-Id: If182c3a6453f7bf8b72f03af56a14ace109eaa68