[WIP] Implement alarm-manager
alarm-manager is responsible for watching a configurable
location within the filesystem where the user can put
a YAML alarm definition file. When change/creation is
detected, the YAML file is checked for proper contents
and if verification is successfull, LUA code is generated
as well as LUA configuration files. Hindsight will pick
up those changes after a certain period of time and
provider proper alarming to the platform

Change-Id: I7b2b98f379c49bdbf23177a038bdca9433d1c6e5