Fix hostname in logs indexed in Elasticsearch
If "hostname" is not set in the global hekad configuration
then Heka uses Go's os.Hostname() as the value of the
Hostname field in log messages. With hekad running in a pod
os.Hostname() returns the pod name, which is not what we
want. This commit fixes it by setting "hostname" in the
global hekad configuration.

Note that this fix relies on the spec.NodeName downward API
variable which is only available in Kubernetes 1.4 and
higher.

Change-Id: Ic7046464e3b73cdde6574cb9b0a591648b7c3b2f