Use base-tools as the base for alarm-manager
In I7b2b98f379c49bdbf23177a038bdca9433d1c6e5 reviews
suggested to use "base" rather than "base-tools" as the base
image. This commit reverts that. Using "base" forces to add
dependencies such as "dumb-init", which should really not be
the responsibility of authors of non-base images.

This commit also changes the way we specify versions of
Python dependencies in requirements.txt. Fixing versions
using `==` allows for more repeatable builds, but it also
means that we will quickly rely on old versions and will not
be able to guarantee that our code works with the latest
versions of our dependencies. Using `>=` is also consistent
with what we currently use in base-tools/requirements.txt.
requirements.txt was updated using openstack_requirement's
update-requirements tool.

Change-Id: I60d0d8e761da717e2f73485bae53f1e11e3aebba