Fix network graphs in Grafana
This commit fixes the network-related graphs in Grafana. The
network statistics provided by Snap are cumulative values.
For example, bytes_recv is the number of bytes received
since the machine boot. For that reason, we need to
configure the Grafana network graphs to use the derivative()
function in SELECT queries.

Change-Id: Icbf5a60c7bc3a34972a4045d454e8ab40917e391